+++ src/data/Data.test.ts
@@ -4,11 +4,11 @@ import { h } from 'react-hyperscript-helpers';
 import { Ajax } from 'src/libs/ajax';
 import { LeoAppStatus, ListAppResponse } from 'src/libs/ajax/leonardo/models/app-models';
 import { WorkspaceWrapper } from 'src/libs/workspace-utils';
-import { WorkspaceData } from 'src/pages/workspaces/workspace/Data';
+import { StorageDetails } from 'src/pages/workspaces/workspace/useWorkspace';
 import { asMockedFn } from 'src/testing/test-utils';
 import { defaultAzureWorkspace, defaultGoogleBucketOptions } from 'src/testing/workspace-fixtures';
 
-import { StorageDetails } from './useWorkspace';
+import { WorkspaceData } from './Data';
 
 type WorkspaceContainerExports = typeof import('src/pages/workspaces/workspace/WorkspaceContainer');
 jest.mock('src/pages/workspaces/workspace/WorkspaceContainer', (): WorkspaceContainerExports => {
